refactor: separate metrics from tracing
